Move getAutomaticDashboardUrl and getComparisonDashboardUrl to metabase/utils/urls
